+++ condition_test.go
@@ -7,6 +7,7 @@ import (
 )
 
 func TestTernary(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	result1 := Ternary(true, "a", "b")
@@ -16,7 +17,18 @@ func TestTernary(t *testing.T) {
 	is.Equal(result2, "b")
 }
 
+func TestTernaryF(t *testing.T) {
+	is := assert.New(t)
+
+	result1 := TernaryF(true, func() string { return "a" }, func() string { return "b" })
+	result2 := TernaryF(false, func() string { return "a" }, func() string { return "b" })
+
+	is.Equal(result1, "a")
+	is.Equal(result2, "b")
+}
+
 func TestIfElse(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	result1 := If(true, 1).ElseIf(false, 2).Else(3)
@@ -31,6 +43,7 @@ func TestIfElse(t *testing.T) {
 }
 
 func TestIfFElseF(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	result1 := IfF(true, func() int { return 1 }).ElseIfF(false, func() int { return 2 }).ElseF(func() int { return 3 })
@@ -45,6 +58,7 @@ func TestIfFElseF(t *testing.T) {
 }
 
 func TestSwitchCase(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	result1 := Switch[int, int](42).Case(42, 1).Case(1, 2).Default(3)
@@ -59,6 +73,7 @@ func TestSwitchCase(t *testing.T) {
 }
 
 func TestSwitchCaseF(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	result1 := Switch[int, int](42).CaseF(42, func() int { return 1 }).CaseF(1, func() int { return 2 }).DefaultF(func() int { return 3 })