+++ map_test.go
@@ -10,6 +10,7 @@ import (
 )
 
 func TestKeys(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	r1 := Keys(map[string]int{"foo": 1, "bar": 2})
@@ -19,6 +20,7 @@ func TestKeys(t *testing.T) {
 }
 
 func TestValues(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	r1 := Values(map[string]int{"foo": 1, "bar": 2})
@@ -28,6 +30,7 @@ func TestValues(t *testing.T) {
 }
 
 func TestPickBy(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	r1 := PickBy(map[string]int{"foo": 1, "bar": 2, "baz": 3}, func(key string, value int) bool {
@@ -38,6 +41,7 @@ func TestPickBy(t *testing.T) {
 }
 
 func TestPickByKeys(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	r1 := PickByKeys(map[string]int{"foo": 1, "bar": 2, "baz": 3}, []string{"foo", "baz"})
@@ -46,6 +50,7 @@ func TestPickByKeys(t *testing.T) {
 }
 
 func TestPickByValues(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	r1 := PickByValues(map[string]int{"foo": 1, "bar": 2, "baz": 3}, []int{1, 3})
@@ -54,6 +59,7 @@ func TestPickByValues(t *testing.T) {
 }
 
 func TestOmitBy(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	r1 := OmitBy(map[string]int{"foo": 1, "bar": 2, "baz": 3}, func(key string, value int) bool {
@@ -64,6 +70,7 @@ func TestOmitBy(t *testing.T) {
 }
 
 func TestOmitByKeys(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	r1 := OmitByKeys(map[string]int{"foo": 1, "bar": 2, "baz": 3}, []string{"foo", "baz"})
@@ -72,6 +79,7 @@ func TestOmitByKeys(t *testing.T) {
 }
 
 func TestOmitByValues(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	r1 := OmitByValues(map[string]int{"foo": 1, "bar": 2, "baz": 3}, []int{1, 3})
@@ -80,6 +88,7 @@ func TestOmitByValues(t *testing.T) {
 }
 
 func TestEntries(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	r1 := Entries(map[string]int{"foo": 1, "bar": 2})
@@ -100,6 +109,7 @@ func TestEntries(t *testing.T) {
 }
 
 func TestFromEntries(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	r1 := FromEntries([]Entry[string, int]{
@@ -119,6 +129,7 @@ func TestFromEntries(t *testing.T) {
 }
 
 func TestInvert(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	r1 := Invert(map[string]int{"a": 1, "b": 2})
@@ -130,6 +141,7 @@ func TestInvert(t *testing.T) {
 }
 
 func TestAssign(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	result1 := Assign(map[string]int{"a": 1, "b": 2}, map[string]int{"b": 3, "c": 4})
@@ -139,6 +151,7 @@ func TestAssign(t *testing.T) {
 }
 
 func TestMapKeys(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	result1 := MapKeys(map[int]int{1: 1, 2: 2, 3: 3, 4: 4}, func(x int, _ int) string {
@@ -154,6 +167,7 @@ func TestMapKeys(t *testing.T) {
 }
 
 func TestMapValues(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	result1 := MapValues(map[int]int{1: 1, 2: 2, 3: 3, 4: 4}, func(x int, _ int) string {
@@ -169,7 +183,92 @@ func TestMapValues(t *testing.T) {
 	is.Equal(result2, map[int]string{1: "1", 2: "2", 3: "3", 4: "4"})
 }
 
+func mapEntriesTest[I any, O any](t *testing.T, in map[string]I, iteratee func(string, I) (string, O), expected map[string]O) {
+	is := assert.New(t)
+	result := MapEntries(in, iteratee)
+	is.Equal(result, expected)
+}
+
+func TestMapEntries(t *testing.T) {
+	mapEntriesTest(t, map[string]int{"foo": 1, "bar": 2}, func(k string, v int) (string, int) {
+		return k, v + 1
+	}, map[string]int{"foo": 2, "bar": 3})
+	mapEntriesTest(t, map[string]int{"foo": 1, "bar": 2}, func(k string, v int) (string, string) {
+		return k, k + strconv.Itoa(v)
+	}, map[string]string{"foo": "foo1", "bar": "bar2"})
+	mapEntriesTest(t, map[string]int{"foo": 1, "bar": 2}, func(k string, v int) (string, string) {
+		return k, strconv.Itoa(v) + k
+	}, map[string]string{"foo": "1foo", "bar": "2bar"})
+
+	// NoMutation
+	{
+		is := assert.New(t)
+		r1 := map[string]int{"foo": 1, "bar": 2}
+		MapEntries(r1, func(k string, v int) (string, string) {
+			return k, strconv.Itoa(v) + "!!"
+		})
+		is.Equal(r1, map[string]int{"foo": 1, "bar": 2})
+	}
+	// EmptyInput
+	{
+		mapEntriesTest(t, map[string]int{}, func(k string, v int) (string, string) {
+			return k, strconv.Itoa(v) + "!!"
+		}, map[string]string{})
+
+		mapEntriesTest(t, map[string]any{}, func(k string, v any) (string, any) {
+			return k, v
+		}, map[string]any{})
+	}
+	// Identity
+	{
+		mapEntriesTest(t, map[string]int{"foo": 1, "bar": 2}, func(k string, v int) (string, int) {
+			return k, v
+		}, map[string]int{"foo": 1, "bar": 2})
+		mapEntriesTest(t, map[string]any{"foo": 1, "bar": "2", "ccc": true}, func(k string, v any) (string, any) {
+			return k, v
+		}, map[string]any{"foo": 1, "bar": "2", "ccc": true})
+	}
+	// ToConstantEntry
+	{
+		mapEntriesTest(t, map[string]any{"foo": 1, "bar": "2", "ccc": true}, func(k string, v any) (string, any) {
+			return "key", "value"
+		}, map[string]any{"key": "value"})
+		mapEntriesTest(t, map[string]any{"foo": 1, "bar": "2", "ccc": true}, func(k string, v any) (string, any) {
+			return "b", 5
+		}, map[string]any{"b": 5})
+	}
+
+	//// OverlappingKeys
+	//// because using range over map, the order is not guaranteed
+	//// this test is not deterministic
+	//{
+	//	mapEntriesTest(t, map[string]any{"foo": 1, "foo2": 2, "Foo": 2, "Foo2": "2", "bar": "2", "ccc": true}, func(k string, v any) (string, any) {
+	//		return string(k[0]), v
+	//	}, map[string]any{"F": "2", "b": "2", "c": true, "f": 2})
+	//	mapEntriesTest(t, map[string]string{"foo": "1", "foo2": "2", "Foo": "2", "Foo2": "2", "bar": "2", "ccc": "true"}, func(k string, v string) (string, string) {
+	//		return v, k
+	//	}, map[string]string{"1": "foo", "2": "bar", "true": "ccc"})
+	//}
+	//NormalMappers
+	{
+		mapEntriesTest(t, map[string]string{"foo": "1", "foo2": "2", "Foo": "2", "Foo2": "2", "bar": "2", "ccc": "true"}, func(k string, v string) (string, string) {
+			return k, k + v
+		}, map[string]string{"Foo": "Foo2", "Foo2": "Foo22", "bar": "bar2", "ccc": "ccctrue", "foo": "foo1", "foo2": "foo22"})
+
+		mapEntriesTest(t, map[string]struct {
+			name string
+			age  int
+		}{"1-11-1": {name: "foo", age: 1}, "2-22-2": {name: "bar", age: 2}}, func(k string, v struct {
+			name string
+			age  int
+		}) (string, string) {
+			return v.name, k
+		}, map[string]string{"bar": "2-22-2", "foo": "1-11-1"})
+	}
+}
+
 func TestMapToSlice(t *testing.T) {
+	t.Parallel()
 	is := assert.New(t)
 
 	result1 := MapToSlice(map[int]int{1: 5, 2: 6, 3: 7, 4: 8}, func(k int, v int) string {