+++ retry.go
@@ -46,6 +46,7 @@ func (d *debounce) cancel() {
 }
 
 // NewDebounce creates a debounced instance that delays invoking functions given until after wait milliseconds have elapsed.
+// Play: https://go.dev/play/p/mz32VMK2nqe
 func NewDebounce(duration time.Duration, f ...func()) (func(), func()) {
 	d := &debounce{
 		after:     duration,
@@ -61,6 +62,7 @@ func NewDebounce(duration time.Duration, f ...func()) (func(), func()) {
 }
 
 // Attempt invokes a function N times until it returns valid output. Returning either the caught error or nil. When first argument is less than `1`, the function runs until a successful response is returned.
+// Play: https://go.dev/play/p/3ggJZ2ZKcMj
 func Attempt(maxIteration int, f func(int) error) (int, error) {
 	var err error
 
@@ -79,6 +81,7 @@ func Attempt(maxIteration int, f func(int) error) (int, error) {
 // with a pause between each call. Returning either the caught error or nil.
 // When first argument is less than `1`, the function runs until a successful
 // response is returned.
+// Play: https://go.dev/play/p/tVs6CygC7m1
 func AttemptWithDelay(maxIteration int, delay time.Duration, f func(int, time.Duration) error) (int, time.Duration, error) {
 	var err error